LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/fill-and-stroke-styles/2d.gradient.radial.cone.bottom-expected.txt

2d.gradient.radial.cone.bottom


FAIL OffscreenCanvas test: 2d.gradient.radial.cone.bottom assert_equals: Green channel of the pixel at (98, 25) expected 255 but got 254
